requirements.txt
Flask==1.0.3
mysql-connector==2.2.9
